Refactor Adapter Config To Its Own File
